ICU-22995 Remove many casts made unnecessary and counterproductive

+++ icu4c/source/test/intltest/convtest.cpp
@@ -816,7 +816,7 @@ ConversionTest::TestUTF8ToUTF8Overflow() {
     assertEquals("illFormed number of bytes written", 0, length);
     errorLength = UPRV_LENGTHOF(errorBytes);
     ucnv_getInvalidChars(cnv1.getAlias(), errorBytes, &errorLength, errorCode);
-    assertEquals("illFormed truncated errorLength", 2, static_cast<int32_t>(errorLength));
+    assertEquals("illFormed truncated errorLength", 2, errorLength);
     if (errorLength == 2) {
         assertEquals("illFormed truncated errorBytes", 0xf191, 
                      (static_cast<int32_t>(static_cast<uint8_t>(errorBytes[0])) << 8) | static_cast<uint8_t>(errorBytes[1]));
@@ -833,7 +833,7 @@ ConversionTest::TestUTF8ToUTF8Overflow() {
     assertEquals("illFormed trail byte number of bytes written", 0, length);
     errorLength = UPRV_LENGTHOF(errorBytes);
     ucnv_getInvalidChars(cnv1.getAlias(), errorBytes, &errorLength, errorCode);
-    assertEquals("illFormed trail byte errorLength", 1, static_cast<int32_t>(errorLength));
+    assertEquals("illFormed trail byte errorLength", 1, errorLength);
     if (errorLength == 1) {
         assertEquals("illFormed trail byte errorBytes", 0x93, static_cast<int32_t>(static_cast<uint8_t>(errorBytes[0])));
     }
